+++ octodns_selectel/escaping_semicolon.py
@@ -1,0 +1,6 @@
+def escape_semicolon(s):
+    return s.replace(';', '\\;')
+
+
+def unescape_semicolon(s):
+    return s.replace('\\;', ';')